refactor: move some common functions to encoding (#24)
* refactor: move some common functions to encoding

* fix golint
