+++ encoding/codecv2/codecv2.go
@@ -1,11 +1,5 @@
 package codecv2
 
-/*
-#include <stdint.h>
-char* compress_scroll_batch_bytes(uint8_t* src, uint64_t src_size, uint8_t* output_buf, uint64_t *output_buf_size);
-*/
-import "C"
-
 import (
 	"crypto/sha256"
 	"encoding/binary"
@@ -13,9 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"unsafe"
 
-	"github.com/scroll-tech/go-ethereum/accounts/abi"
 	"github.com/scroll-tech/go-ethereum/common"
 	"github.com/scroll-tech/go-ethereum/core/types"
 	"github.com/scroll-tech/go-ethereum/crypto"
@@ -24,6 +16,7 @@ import (
 
 	"github.com/scroll-tech/da-codec/encoding"
 	"github.com/scroll-tech/da-codec/encoding/codecv1"
+	"github.com/scroll-tech/da-codec/encoding/zstd"
 )
 
 // MaxNumChunks is the maximum number of chunks that a batch can contain.
@@ -176,7 +169,7 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, useMockTxData bool) (*kzg484
 	copy(challengePreimage[0:], hash[:])
 
 	// blobBytes represents the compressed blob payload (batchBytes)
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return nil, common.Hash{}, nil, nil, err
 	}
@@ -196,7 +189,7 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, useMockTxData bool) (*kzg484
 	}
 
 	// convert raw data to BLSFieldElements
-	blob, err := MakeBlobCanonical(blobBytes)
+	blob, err := encoding.MakeBlobCanonical(blobBytes)
 	if err != nil {
 		return nil, common.Hash{}, nil, nil, err
 	}
@@ -222,11 +215,6 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, useMockTxData bool) (*kzg484
 	copy(z[start:], pointBytes)
 
 	return blob, blobVersionedHash, &z, blobBytes, nil
-}
-
-// MakeBlobCanonical converts the raw blob data into the canonical blob representation of 4096 BLSFieldElements.
-func MakeBlobCanonical(blobBytes []byte) (*kzg4844.Blob, error) {
-	return codecv1.MakeBlobCanonical(blobBytes)
 }
 
 // NewDABatchFromBytes decodes the given byte slice into a DABatch.
@@ -295,7 +283,7 @@ func (b *DABatch) BlobDataProof() ([]byte, error) {
 	// | bytes32 | bytes32 | bytes48        | bytes48   |
 
 	values := []interface{}{*b.z, y, commitment, proof}
-	blobDataProofArgs, err := GetBlobDataProofArgs()
+	blobDataProofArgs, err := encoding.GetBlobDataProofArgs()
 	if err != nil {
 		return nil, fmt.Errorf("failed to get blob data proof args, err: %w", err)
 	}
@@ -309,38 +297,38 @@ func (b *DABatch) Blob() *kzg4844.Blob {
 
 // EstimateChunkL1CommitBatchSizeAndBlobSize estimates the L1 commit uncompressed batch size and compressed blob size for a single chunk.
 func EstimateChunkL1CommitBatchSizeAndBlobSize(c *encoding.Chunk) (uint64, uint64, error) {
-	batchBytes, err := constructBatchPayload([]*encoding.Chunk{c})
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob([]*encoding.Chunk{c}, MaxNumChunks)
 	if err != nil {
 		return 0, 0, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return 0, 0, err
 	}
-	return uint64(len(batchBytes)), CalculatePaddedBlobSize(uint64(len(blobBytes))), nil
+	return uint64(len(batchBytes)), encoding.CalculatePaddedBlobSize(uint64(len(blobBytes))), nil
 }
 
 // EstimateBatchL1CommitBatchSizeAndBlobSize estimates the L1 commit uncompressed batch size and compressed blob size for a batch.
 func EstimateBatchL1CommitBatchSizeAndBlobSize(b *encoding.Batch) (uint64, uint64, error) {
-	batchBytes, err := constructBatchPayload(b.Chunks)
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob(b.Chunks, MaxNumChunks)
 	if err != nil {
 		return 0, 0, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return 0, 0, err
 	}
-	return uint64(len(batchBytes)), CalculatePaddedBlobSize(uint64(len(blobBytes))), nil
+	return uint64(len(batchBytes)), encoding.CalculatePaddedBlobSize(uint64(len(blobBytes))), nil
 }
 
 // CheckChunkCompressedDataCompatibility checks the compressed data compatibility for a batch built from a single chunk.
 // It constructs a batch payload, compresses the data, and checks the compressed data compatibility if the uncompressed data exceeds 128 KiB.
 func CheckChunkCompressedDataCompatibility(c *encoding.Chunk) (bool, error) {
-	batchBytes, err := constructBatchPayload([]*encoding.Chunk{c})
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob([]*encoding.Chunk{c}, MaxNumChunks)
 	if err != nil {
 		return false, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return false, err
 	}
@@ -358,11 +346,11 @@ func CheckChunkCompressedDataCompatibility(c *encoding.Chunk) (bool, error) {
 // CheckBatchCompressedDataCompatibility checks the compressed data compatibility for a batch.
 // It constructs a batch payload, compresses the data, and checks the compressed data compatibility if the uncompressed data exceeds 128 KiB.
 func CheckBatchCompressedDataCompatibility(b *encoding.Batch) (bool, error) {
-	batchBytes, err := constructBatchPayload(b.Chunks)
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob(b.Chunks, MaxNumChunks)
 	if err != nil {
 		return false, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return false, err
 	}
@@ -400,69 +388,4 @@ func EstimateChunkL1CommitGas(c *encoding.Chunk) uint64 {
 // EstimateBatchL1CommitGas calculates the total L1 commit gas for this batch approximately.
 func EstimateBatchL1CommitGas(b *encoding.Batch) uint64 {
 	return codecv1.EstimateBatchL1CommitGas(b)
-}
-
-// constructBatchPayload constructs the batch payload.
-// This function is only used in compressed batch payload length estimation.
-func constructBatchPayload(chunks []*encoding.Chunk) ([]byte, error) {
-	// metadata consists of num_chunks (2 bytes) and chunki_size (4 bytes per chunk)
-	metadataLength := 2 + MaxNumChunks*4
-
-	// batchBytes represents the raw (un-compressed and un-padded) blob payload
-	batchBytes := make([]byte, metadataLength)
-
-	// batch metadata: num_chunks
-	binary.BigEndian.PutUint16(batchBytes[0:], uint16(len(chunks)))
-
-	// encode batch metadata and L2 transactions,
-	for chunkID, chunk := range chunks {
-		currentChunkStartIndex := len(batchBytes)
-
-		for _, block := range chunk.Blocks {
-			for _, tx := range block.Transactions {
-				if tx.Type == types.L1MessageTxType {
-					continue
-				}
-
-				// encode L2 txs into batch payload
-				rlpTxData, err := encoding.ConvertTxDataToRLPEncoding(tx, false /* no mock */)
-				if err != nil {
-					return nil, err
-				}
-				batchBytes = append(batchBytes, rlpTxData...)
-			}
-		}
-
-		// batch metadata: chunki_size
-		if chunkSize := len(batchBytes) - currentChunkStartIndex; chunkSize != 0 {
-			binary.BigEndian.PutUint32(batchBytes[2+4*chunkID:], uint32(chunkSize))
-		}
-	}
-	return batchBytes, nil
-}
-
-// compressScrollBatchBytes compresses the given batch of bytes.
-// The output buffer is allocated with an extra 128 bytes to accommodate metadata overhead or error message.
-func compressScrollBatchBytes(batchBytes []byte) ([]byte, error) {
-	srcSize := C.uint64_t(len(batchBytes))
-	outbufSize := C.uint64_t(len(batchBytes) + 128) // Allocate output buffer with extra 128 bytes
-	outbuf := make([]byte, outbufSize)
-
-	if err := C.compress_scroll_batch_bytes((*C.uchar)(unsafe.Pointer(&batchBytes[0])), srcSize,
-		(*C.uchar)(unsafe.Pointer(&outbuf[0])), &outbufSize); err != nil {
-		return nil, fmt.Errorf("failed to compress scroll batch bytes: %s", C.GoString(err))
-	}
-
-	return outbuf[:int(outbufSize)], nil
-}
-
-// CalculatePaddedBlobSize calculates the required size on blob storage
-// where every 32 bytes can store only 31 bytes of actual data, with the first byte being zero.
-func CalculatePaddedBlobSize(dataSize uint64) uint64 {
-	return codecv1.CalculatePaddedBlobSize(dataSize)
-}
-
-// GetBlobDataProofArgs gets the blob data proof arguments for batch commitment and returns error if initialization fails.
-func GetBlobDataProofArgs() (*abi.Arguments, error) {
-	return codecv1.GetBlobDataProofArgs()
 }
