+++ encoding/codecv4/codecv4.go
@@ -1,11 +1,5 @@
 package codecv4
 
-/*
-#include <stdint.h>
-char* compress_scroll_batch_bytes(uint8_t* src, uint64_t src_size, uint8_t* output_buf, uint64_t *output_buf_size);
-*/
-import "C"
-
 import (
 	"crypto/sha256"
 	"encoding/binary"
@@ -13,9 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"unsafe"
 
-	"github.com/scroll-tech/go-ethereum/accounts/abi"
 	"github.com/scroll-tech/go-ethereum/common"
 	"github.com/scroll-tech/go-ethereum/core/types"
 	"github.com/scroll-tech/go-ethereum/crypto"
@@ -23,8 +15,8 @@ import (
 	"github.com/scroll-tech/go-ethereum/log"
 
 	"github.com/scroll-tech/da-codec/encoding"
-	"github.com/scroll-tech/da-codec/encoding/codecv1"
 	"github.com/scroll-tech/da-codec/encoding/codecv3"
+	"github.com/scroll-tech/da-codec/encoding/zstd"
 )
 
 // MaxNumChunks is the maximum number of chunks that a batch can contain.
@@ -197,7 +189,7 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, enableCompress bool, useMock
 	if enableCompress {
 		// blobBytes represents the compressed blob payload (batchBytes)
 		var err error
-		blobBytes, err = compressScrollBatchBytes(batchBytes)
+		blobBytes, err = zstd.CompressScrollBatchBytes(batchBytes)
 		if err != nil {
 			return nil, common.Hash{}, nil, nil, err
 		}
@@ -210,7 +202,6 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, enableCompress bool, useMock
 		}
 		blobBytes = append([]byte{1}, blobBytes...)
 	} else {
-		blobBytes = batchBytes
 		blobBytes = append([]byte{0}, batchBytes...)
 	}
 
@@ -220,7 +211,7 @@ func ConstructBlobPayload(chunks []*encoding.Chunk, enableCompress bool, useMock
 	}
 
 	// convert raw data to BLSFieldElements
-	blob, err := MakeBlobCanonical(blobBytes)
+	blob, err := encoding.MakeBlobCanonical(blobBytes)
 	if err != nil {
 		return nil, common.Hash{}, nil, nil, err
 	}
@@ -345,7 +336,7 @@ func (b *DABatch) BlobDataProofForPointEvaluation() ([]byte, error) {
 	// | bytes32 | bytes32 | bytes48        | bytes48   |
 
 	values := []interface{}{*b.z, y, commitment, proof}
-	blobDataProofArgs, err := GetBlobDataProofArgs()
+	blobDataProofArgs, err := encoding.GetBlobDataProofArgs()
 	if err != nil {
 		return nil, fmt.Errorf("failed to get blob data proof args, err: %w", err)
 	}
@@ -364,13 +355,13 @@ func (b *DABatch) BlobBytes() []byte {
 
 // EstimateChunkL1CommitBatchSizeAndBlobSize estimates the L1 commit uncompressed batch size and compressed blob size for a single chunk.
 func EstimateChunkL1CommitBatchSizeAndBlobSize(c *encoding.Chunk, enableCompress bool) (uint64, uint64, error) {
-	batchBytes, err := constructBatchPayload([]*encoding.Chunk{c})
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob([]*encoding.Chunk{c}, MaxNumChunks)
 	if err != nil {
 		return 0, 0, err
 	}
 	var blobBytesLength uint64
 	if enableCompress {
-		blobBytes, err := compressScrollBatchBytes(batchBytes)
+		blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 		if err != nil {
 			return 0, 0, err
 		}
@@ -378,18 +369,18 @@ func EstimateChunkL1CommitBatchSizeAndBlobSize(c *encoding.Chunk, enableCompress
 	} else {
 		blobBytesLength = 1 + uint64(len(batchBytes))
 	}
-	return uint64(len(batchBytes)), CalculatePaddedBlobSize(blobBytesLength), nil
+	return uint64(len(batchBytes)), encoding.CalculatePaddedBlobSize(blobBytesLength), nil
 }
 
 // EstimateBatchL1CommitBatchSizeAndBlobSize estimates the L1 commit uncompressed batch size and compressed blob size for a batch.
 func EstimateBatchL1CommitBatchSizeAndBlobSize(b *encoding.Batch, enableCompress bool) (uint64, uint64, error) {
-	batchBytes, err := constructBatchPayload(b.Chunks)
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob(b.Chunks, MaxNumChunks)
 	if err != nil {
 		return 0, 0, err
 	}
 	var blobBytesLength uint64
 	if enableCompress {
-		blobBytes, err := compressScrollBatchBytes(batchBytes)
+		blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 		if err != nil {
 			return 0, 0, err
 		}
@@ -397,16 +388,16 @@ func EstimateBatchL1CommitBatchSizeAndBlobSize(b *encoding.Batch, enableCompress
 	} else {
 		blobBytesLength = 1 + uint64(len(batchBytes))
 	}
-	return uint64(len(batchBytes)), CalculatePaddedBlobSize(blobBytesLength), nil
+	return uint64(len(batchBytes)), encoding.CalculatePaddedBlobSize(blobBytesLength), nil
 }
 
 // CheckChunkCompressedDataCompatibility checks the compressed data compatibility for a batch built from a single chunk.
 func CheckChunkCompressedDataCompatibility(c *encoding.Chunk) (bool, error) {
-	batchBytes, err := constructBatchPayload([]*encoding.Chunk{c})
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob([]*encoding.Chunk{c}, MaxNumChunks)
 	if err != nil {
 		return false, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return false, err
 	}
@@ -419,11 +410,11 @@ func CheckChunkCompressedDataCompatibility(c *encoding.Chunk) (bool, error) {
 
 // CheckBatchCompressedDataCompatibility checks the compressed data compatibility for a batch.
 func CheckBatchCompressedDataCompatibility(b *encoding.Batch) (bool, error) {
-	batchBytes, err := constructBatchPayload(b.Chunks)
+	batchBytes, err := encoding.ConstructBatchPayloadInBlob(b.Chunks, MaxNumChunks)
 	if err != nil {
 		return false, err
 	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
+	blobBytes, err := zstd.CompressScrollBatchBytes(batchBytes)
 	if err != nil {
 		return false, err
 	}
@@ -452,92 +443,4 @@ func EstimateChunkL1CommitGas(c *encoding.Chunk) uint64 {
 // EstimateBatchL1CommitGas calculates the total L1 commit gas for this batch approximately.
 func EstimateBatchL1CommitGas(b *encoding.Batch) uint64 {
 	return codecv3.EstimateBatchL1CommitGas(b)
-}
-
-// GetBlobDataProofArgs gets the blob data proof arguments for batch commitment and returns error if initialization fails.
-func GetBlobDataProofArgs() (*abi.Arguments, error) {
-	return codecv3.GetBlobDataProofArgs()
-}
-
-// checkBatchCompressedDataCompatibility checks the compressed data compatibility for a batch.
-// It constructs a batch payload, compresses the data, and checks the compressed data compatibility if the uncompressed data exceeds 128 KiB.
-func checkBatchCompressedDataCompatibility(b *encoding.Batch) (bool, error) {
-	batchBytes, err := constructBatchPayload(b.Chunks)
-	if err != nil {
-		return false, err
-	}
-	blobBytes, err := compressScrollBatchBytes(batchBytes)
-	if err != nil {
-		return false, err
-	}
-	if err = encoding.CheckCompressedDataCompatibility(blobBytes); err != nil {
-		log.Warn("CheckBatchCompressedDataCompatibility: compressed data compatibility check failed", "err", err, "batchBytes", hex.EncodeToString(batchBytes), "blobBytes", hex.EncodeToString(blobBytes))
-		return false, nil
-	}
-	return true, nil
-}
-
-// constructBatchPayload constructs the batch payload.
-// This function is only used in compressed batch payload length estimation.
-func constructBatchPayload(chunks []*encoding.Chunk) ([]byte, error) {
-	// metadata consists of num_chunks (2 bytes) and chunki_size (4 bytes per chunk)
-	metadataLength := 2 + MaxNumChunks*4
-
-	// batchBytes represents the raw (un-compressed and un-padded) blob payload
-	batchBytes := make([]byte, metadataLength)
-
-	// batch metadata: num_chunks
-	binary.BigEndian.PutUint16(batchBytes[0:], uint16(len(chunks)))
-
-	// encode batch metadata and L2 transactions,
-	for chunkID, chunk := range chunks {
-		currentChunkStartIndex := len(batchBytes)
-
-		for _, block := range chunk.Blocks {
-			for _, tx := range block.Transactions {
-				if tx.Type == types.L1MessageTxType {
-					continue
-				}
-
-				// encode L2 txs into batch payload
-				rlpTxData, err := encoding.ConvertTxDataToRLPEncoding(tx, false /* no mock */)
-				if err != nil {
-					return nil, err
-				}
-				batchBytes = append(batchBytes, rlpTxData...)
-			}
-		}
-
-		// batch metadata: chunki_size
-		if chunkSize := len(batchBytes) - currentChunkStartIndex; chunkSize != 0 {
-			binary.BigEndian.PutUint32(batchBytes[2+4*chunkID:], uint32(chunkSize))
-		}
-	}
-	return batchBytes, nil
-}
-
-// compressScrollBatchBytes compresses the given batch of bytes.
-// The output buffer is allocated with an extra 128 bytes to accommodate metadata overhead or error message.
-func compressScrollBatchBytes(batchBytes []byte) ([]byte, error) {
-	srcSize := C.uint64_t(len(batchBytes))
-	outbufSize := C.uint64_t(len(batchBytes) + 128) // Allocate output buffer with extra 128 bytes
-	outbuf := make([]byte, outbufSize)
-
-	if err := C.compress_scroll_batch_bytes((*C.uchar)(unsafe.Pointer(&batchBytes[0])), srcSize,
-		(*C.uchar)(unsafe.Pointer(&outbuf[0])), &outbufSize); err != nil {
-		return nil, fmt.Errorf("failed to compress scroll batch bytes: %s", C.GoString(err))
-	}
-
-	return outbuf[:int(outbufSize)], nil
-}
-
-// MakeBlobCanonical converts the raw blob data into the canonical blob representation of 4096 BLSFieldElements.
-func MakeBlobCanonical(blobBytes []byte) (*kzg4844.Blob, error) {
-	return codecv1.MakeBlobCanonical(blobBytes)
-}
-
-// CalculatePaddedBlobSize calculates the required size on blob storage
-// where every 32 bytes can store only 31 bytes of actual data, with the first byte being zero.
-func CalculatePaddedBlobSize(dataSize uint64) uint64 {
-	return codecv1.CalculatePaddedBlobSize(dataSize)
 }
