+++ encoding/zstd/libscroll_zstd_darwin_arm64.go
@@ -1,4 +1,4 @@
-package codecv2
+package zstd
 
 /*
 #cgo LDFLAGS: ${SRCDIR}/libscroll_zstd_darwin_arm64.a
